Move numpy_quantile() and type check differently for python 3.7 / 3.8+
