+++ great_expectations/rule_based_profiler/helpers/util.py
@@ -59,8 +59,8 @@ from great_expectations.util import (
     convert_ndarray_datetime_to_float_dtype_utc_timezone,
     convert_ndarray_float_to_datetime_dtype,
     convert_ndarray_to_datetime_dtype_best_effort,
-    numpy_quantile,
 )
+from great_expectations.compatibility.numpy import numpy_quantile
 from great_expectations.validator.computed_metric import MetricValue  # noqa: TCH001
 from great_expectations.validator.metric_configuration import (
     MetricConfiguration,  # noqa: TCH001
